CMakeLists.txt: Allow pseudo-static build on Ubuntu
Allow STATIC_LINK=ON on Ubuntu by:

    * providing libebl
    * avoiding a trailing -Wl,-Bstatic that makes the resulting binary
      segfault
    * dynamically link to libdl. Static linking to that lib would impose
      that the same glibc version used for linking is present on the
      runtime system, which is not very safe and would mislead into
      thinking the binary is safe to use everywhere.

Apart from that, all the other libraries are statically linked to the
executable.

Note: the user might need a symlink /lib/ld64.so.1 => /lib64/ld-linux-x86-64.so.2
The ld option --dynamic-linker=file is supposed to allow fixing that,
but it's unfortunately ignored when used with -static.