fprintf: Handle anonymous enum
Ensure anonymous enumerations can be printed multiple types. It is
annoyingly necessary, as an enumeration:

  1. Cannot be empty
  2. Cannot be redefined with the same constants

This means that multiple uses of the same anonymous enumeration (e.g.
when deep-printing a containing struct type) have no way of referring to
the previous definition (since it's anonymous), and also cannot simply
redefine the enumeration inline without changes (as this would create a
re-definition error).

Fix that by adding a pahole-specific prefix to each member for the follow-up
redefinitions. This also ensures that the selected integral type will be
the same between all the enums, as they will be defined using the same
list of constant values (just with different names).

Signed-off-by: Douglas Raillard <douglas.raillard@arm.com>